+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+* Added: New function Connect-SqlServer
+
 ## 0.4.2 - 2020-07-14
 
 * Fixed: No output of the last state for the transaction log shrinking